Integrate node/src/loader.rs into node/src/lib.rs
Make things simple for refactoring - merge all related
parts into one file.
